Moving classes into separate files
