+++ src/DrawingLink.UI/Enums.cs
@@ -1,0 +1,17 @@
+﻿namespace DrawingLink.UI
+{
+    public enum ModelParamType
+    {
+        Point,
+        Line,
+        Polyline,
+        Face,
+        Object
+    }
+
+    public enum DrawingParamType
+    {
+        Point,
+        Object
+    }
+}
