+++ src/DrawingLink.UI/GH/ActiveObjectWrapper.cs
@@ -1,0 +1,21 @@
+﻿using Grasshopper.Kernel;
+using System;
+
+namespace DrawingLink.UI.GH
+{
+    public class ActiveObjectWrapper
+    {
+        public IGH_ActiveObject ActiveObject { get; }
+        public TabAndGroup TabAndGroup { get; }
+        public ObjectConnectivity Connectivity { get; }
+        public string FieldName { get; set; }
+
+        public ActiveObjectWrapper(IGH_ActiveObject activeObject, TabAndGroup tabAndGroup, ObjectConnectivity connectivity)
+        {
+            ActiveObject = activeObject ?? throw new ArgumentNullException(nameof(activeObject));
+            TabAndGroup = tabAndGroup ?? throw new ArgumentNullException(nameof(tabAndGroup));
+            Connectivity = connectivity ?? throw new ArgumentNullException(nameof(connectivity));
+            FieldName = string.Empty;
+        }
+    }
+}
