+++ src/DrawingLink.UI/GH/GHParams.cs
@@ -4,7 +4,6 @@ using Grasshopper.Kernel.Special;
 using Grasshopper.Kernel.Types;
 using System;
 using System.Collections.Generic;
-using TSG = Tekla.Structures.Geometry3d;
 
 namespace DrawingLink.UI.GH
 {
@@ -66,163 +65,5 @@ namespace DrawingLink.UI.GH
             foreach (var param in TeklaParams.DrawingParams)
                 param.FieldName = $"tekla_{counter++}";
         }
-    }
-
-    public class ActiveObjectWrapper
-    {
-        public IGH_ActiveObject ActiveObject { get; }
-        public TabAndGroup TabAndGroup { get; }
-        public ObjectConnectivity Connectivity { get; }
-        public string FieldName { get; set; }
-
-        public ActiveObjectWrapper(IGH_ActiveObject activeObject, TabAndGroup tabAndGroup, ObjectConnectivity connectivity)
-        {
-            ActiveObject = activeObject ?? throw new ArgumentNullException(nameof(activeObject));
-            TabAndGroup = tabAndGroup ?? throw new ArgumentNullException(nameof(tabAndGroup));
-            Connectivity = connectivity ?? throw new ArgumentNullException(nameof(connectivity));
-            FieldName = string.Empty;
-        }
-    }
-
-    public abstract class TeklaParamBase
-    {
-        public IGH_ActiveObject ActiveObject { get; }
-        public string FieldName { get; set; }
-        public bool IsMultiple { get; }
-        public string Prompt { get; }
-
-        public TeklaObjects TeklaObjects { get; protected set; }
-
-        protected TeklaParamBase(IGH_ActiveObject activeObject, bool isMultiple, string prompt)
-        {
-            ActiveObject = activeObject ?? throw new ArgumentNullException(nameof(activeObject));
-            IsMultiple = isMultiple;
-            Prompt = prompt ?? throw new ArgumentNullException(nameof(prompt));
-            FieldName = string.Empty;
-        }
-    }
-
-    public class TeklaModelParam : TeklaParamBase
-    {
-        public ModelParamType ParamType { get; }
-
-        public TeklaModelParam(IGH_ActiveObject activeObject, ModelParamType paramType, bool isMultiple, string prompt) : base(activeObject, isMultiple, prompt)
-        {
-            ParamType = paramType;
-
-            TeklaObjects = new TeklaObjects(isDrawing: false);
-            TeklaObjects.SetModelType(ParamType);
-        }
-
-        public void Set(List<Tekla.Structures.Model.ModelObject> modelObjects)
-        {
-            TeklaObjects.Set(modelObjects);
-        }
-
-        public void Set(List<Tekla.Structures.Geometry3d.Point> points)
-        {
-            TeklaObjects.Set(points);
-        }
-    }
-
-    public class TeklaObjects
-    {
-        private List<Tekla.Structures.Model.ModelObject>? _modelObjects;
-        private List<Tekla.Structures.Drawing.DatabaseObject>? _drawingObjects;
-        private List<Tekla.Structures.Geometry3d.Point>? _points;
-
-        private readonly bool _isDrawing;
-        private DrawingParamType _drawingType;
-        private ModelParamType _modelType;
-
-        public TeklaObjects(bool isDrawing)
-        {
-            _isDrawing = isDrawing;
-        }
-
-        public void SetDrawingType(DrawingParamType drawingParamType)
-        {
-            _drawingType = drawingParamType;
-        }
-
-        public void SetModelType(ModelParamType modelParamType)
-        {
-            _modelType = modelParamType;
-        }
-
-        public object[] GetCorrectObject()
-        {
-            if (_isDrawing)
-            {
-                return _drawingType == DrawingParamType.Point ? _points.ToArray() : _drawingObjects.ToArray();
-            }
-            else
-            {
-                return _modelType switch
-                {
-                    ModelParamType.Point => _points.ToArray(),
-                    ModelParamType.Line => new object[] { new TSG.LineSegment(_points[0], _points[1]) },
-                    ModelParamType.Polyline => new object[] { new TSG.PolyLine(_points) },
-                    ModelParamType.Face => new object[] { new TSG.PolyLine(_points) },
-                    _ => _modelObjects.ToArray(),
-                };
-            }
-        }
-
-        internal void Set(List<Tekla.Structures.Model.ModelObject> modelObjects)
-        {
-            _modelObjects = modelObjects;
-        }
-
-        public void Set(List<Tekla.Structures.Geometry3d.Point> points)
-        {
-            _points = points;
-        }
-
-        public void Set(List<Tekla.Structures.Drawing.DatabaseObject> drawingObjects)
-        {
-            _drawingObjects = drawingObjects;
-        }
-    }
-
-    public class TeklaDrawingParam : TeklaParamBase
-    {
-        public DrawingParamType ParamType { get; }
-
-        private List<Tekla.Structures.Drawing.DatabaseObject>? _drawingObjects;
-        private List<Tekla.Structures.Geometry3d.Point>? _points;
-
-        public TeklaDrawingParam(IGH_ActiveObject activeObject, DrawingParamType paramType, bool isMultiple, string prompt) : base(activeObject, isMultiple, prompt)
-        {
-            ParamType = paramType;
-
-            TeklaObjects = new TeklaObjects(isDrawing: true);
-            TeklaObjects.SetDrawingType(ParamType);
-        }
-
-        public void Set(List<Tekla.Structures.Drawing.DatabaseObject> drawingObjects)
-        {
-            TeklaObjects.Set(drawingObjects);
-        }
-
-        public void Set(List<Tekla.Structures.Geometry3d.Point> points)
-        {
-            TeklaObjects.Set(points);
-        }
-    }
-
-    public enum ModelParamType
-    {
-        Point,
-        Line,
-        Polyline,
-        Face,
-        Object
-    }
-
-    public enum DrawingParamType
-    {
-        Point,
-        Object
     }
 }
